mro_package_moved only needs one loop
We can avoid the double iteration by doing the first iteration’s job
(to clear the linearisations) inside gather_and_rename as the items
are added to the stash. The important thing is that they all be
cleared before *any* calls to mro_isa_changed_in, which will still
be the case.

Maybe ‘gather_and_rename’ is not such a good name any more....