In Perl_pad_new(), avoid calling av_fetch() for something we already know.
We've just stored padname and pad as the first two elements in padlist - so
calling av_fetch() is makework.

The code was the way it was because dd2155a49b710f23 moved the two halves from
disjoint locations into the same function adjacent to each other, but didn't
spot this subsequent optimisation.