In Perl_pad_new(), allocate a 2 element array for padlist.
Most subroutines never recurse, hence only need 2 entries in the padlist
array - names, and depth=1.  The default for av_store() is to allocate 0..3,
and even an explicit call to av_extend() with <3 will be rounded up, so we
inline the allocation of the array here.

Running ./installman allocates 7K less with this change.