Work-around Uni 6.0 issues with 'BELL'
Unicode version 6.0 has co-opted the name BELL for a different character
than traditionally used in Perl.  This patch works around that by adding
ALERT as a synonym for BELL, and causing a deprecated warning for uses
of the old name.

The new Unicode character will be nameless in Perl 5.14, unless I can
(unlikely) get Unicode to grant a synonym that they will support.