UCD.pm: Don't use NamedSequences.txt, saves disk
This changes UCD to not use this file.  Instead it takes advantage of
the recent addition of named sequences being accessible through the \N{}
construct.  In one case where it returns a hash of all the named
sequences, it uses the same .pl file that \N{} does.  My guess is that
this routine's usefulness is now past, as named sequences are now
incorporated into the core.