mktables: Smaller tables if not -DDEBUGGING
This patch frees up about .5Mb of disk space.  First, it omits an
unnecessary trailing tab in the output tables, at the cost of an extra
test.  Secondly, for non-debugging Perls, it omits the
helpful-to-developer comments at the beginning of each table file, and
changes the default to not output the other helpful-to-developer
comments that give the number of code points in each output range in the
tables.