[perl #77926] Glob reification during localisation
This was supposed to have been fixed by 2acc3314e31a9, but the code it
added was in the wrong spot. (This is the code that makes *{} return a
new glob if its argument is FAKE.)

Consequently, local *$foo was localising $foo, not *$foo.

This changes the order of the statements and adds a test.