Fix perl build problems on Stratus VOS
This patch is a port of commit a5addb167c102dc5dcd1ab886caf0cb4f554eb05 to
the 5.12 maintenance branch.  It removes the extra "miniperl" dependency
of the "all" target in Makefile.SH, and also tweaks ext/Socket/Socket.xs
to build on systems that do not have the relevant IPv6 defines.