regcomp.c: Fix longjmp-related warnings
This patch should get rid of the compiler warnings recently introduced.

Another way to handle the pm_flags warning is to declare it to be
volatile, but not all compilers that perl uses apparently have that, so
I chose a longer way of introducting a new variable that isn't changed
within the jumpable area.  The others were fixed by not initializing
them before the jumpable area.