regcomp.c: Remove no longer necessary loop
Recent changes to this cause the bitmap to be populated where possible
with the all folding taken into consideration.  Therefore, the FOLD flag
isn't necessary except when that wasn't possible, and the loop that went
through looking for folds is also not necessary, as the bitmap is
now completely populated before it gets to where the loop was.