Always build with crypt() support on Windows
The sources did support to drop win32/fcrypt.c from a Perl distribution
and build without a crypt() implementation to account for the paranoia
around distribution of encryption technology.  However sources and binaries
have been released for at least 10 years with all the code in place, so
there is not much point in making it configurable.